Move pruning SQL queries to a middleware
