+++ packages/playground/sync/src/middleware/utils.ts
@@ -15,3 +15,20 @@ export function asSQLMapper(
 			return message;
 		});
 }
+
+export function asSQLFlatMapper(
+	flatMapper: (message: SQLQueryMetadata) => SQLQueryMetadata[]
+) {
+	return (messages: TransportMessage[]) =>
+		messages
+			.map((message) => {
+				if (message.scope === 'sql') {
+					return {
+						...message,
+						details: message.details.flatMap(flatMapper),
+					};
+				}
+				return message;
+			})
+			.filter((message) => message.details.length > 0);
+}
